Refactor Note Service.
